[FIXED JENKINS-20599] Modified TAPFileNameResultSeeker to AbstractTAPFileNameResultSeeker in order to make it an abstract class and extend it with a new TAPFileNameResultSeeker and TAPMultiTestpointsFileNameResultSeeker
